major update

+++ src/main/java/com/alex/d/springbootatm/controller/ManagerController.java
@@ -3,8 +3,11 @@ package com.alex.d.springbootatm.controller;
 import com.alex.d.springbootatm.exception.CardNotFoundException;
 import com.alex.d.springbootatm.model.BankCard;
 import com.alex.d.springbootatm.repository.BankCardRepository;
+import com.alex.d.springbootatm.response.ErrorResponse;
 import com.alex.d.springbootatm.service.ATMService;
 import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.extern.slf4j.Slf4j;
@@ -12,6 +15,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.time.Instant;
 import java.util.List;
 
 @Slf4j
@@ -47,18 +51,25 @@ public class ManagerController {
             description = "Delete all details about card",
             responses = {
                     @ApiResponse(responseCode = "200", description = "Card was deleted."),
-                    @ApiResponse(responseCode = "404", description = "Failed to delete, card not found ")
+                    @ApiResponse(responseCode = "404", description = "Invalid credit card number", content = {
+                            @Content(mediaType = "application/json;charset=UTF-8", schema = @Schema(implementation = ErrorResponse.class))}),
             }
     )
-
     @DeleteMapping("/delete/{cardNumber}")
-    public ResponseEntity<Void> deleteCard (@PathVariable("cardNumber") String cardNumber) {
+    public ResponseEntity deleteCard(@PathVariable("cardNumber") String cardNumber) {
         try {
+            BankCard recipientCard = bankCardRepository.findByCardNumber(cardNumber);
+            if (recipientCard == null) {
+                log.error("Invalid credit card number {}", cardNumber);
+                ErrorResponse errorResponse = new ErrorResponse(Instant.now(),404, "Card not found", "/delete/" + cardNumber);
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body(errorResponse);
+            }
+
             atmService.deleteCardByNumber(cardNumber);
             log.info("Card with number {} was deleted", cardNumber);
             return ResponseEntity.ok().build();
         } catch (CardNotFoundException e) {
-            log.error("Failed to delete card: {}", e.getMessage(), new CardNotFoundException("Card not found exception."));
+            log.error("Failed to delete card: {}", e.getMessage());
             return ResponseEntity.notFound().build();
         }
     }
@@ -77,7 +88,6 @@ public class ManagerController {
         log.info("New card created: {}", createdCard.getCardNumber());
         return new ResponseEntity<>(createdCard, HttpStatus.CREATED);
     }
-
 
 
 }
